More concise recommended list
Problem: it's easy with the current "show four most recent" to run into
situations where- for example- only GHC 7.8 has had a successful build in
the past few days, and therefore it's the only snapshot that appears.
This gives a more consistent and useful listing of recommended
snapshots.

Styling of this is up for debate.