Temporary switch to new k8s secrets name
This is needed to we can easily migrate to the new db with the
recent pantry changes. If something goes wrong we can simply revert to
the old secret (and hence the old db).

Once the upgrade is done (and verified) we can update the old secret and
update the name in this commit (back to it's original name)