Allow customizing badge color with query string
With this change, the /package/#PackageNameP/badge/#SnapshotBranch/
endpoint takes a new query string "color" that controls the badge color.
The color can be specified as a hex code or a named color from
Graphics.Badge.Barrier.Color in the barrier library. The badge color
defaults to "brightgreen", preserving default behavior when no query
string is supplied.